Fix misc small bugs in game logic
Add functionality to reset ship properties, such as the number of
hits, when a new game is started.
Add parameter to gameboard.printBoard method to hide ships so computer
board ships are not shown to the user.
Throw error to prevent a single unit length ship to be created
inside the ship factory function.